Corrections in code styling for Samplers. Some corrections in ProteinEDAOptimizer --> still have to remove debug messages.

+++ projects/Proteins/Rosetta/Sampler/EnumerationDiscreteSampler.h
@@ -21,32 +21,32 @@ namespace lbcpp
 class EnumerationDiscreteSampler;
 typedef ReferenceCountedObjectPtr<EnumerationDiscreteSampler> EnumerationDiscreteSamplerPtr;
 
-class EnumerationDiscreteSampler: public DiscreteSampler
+class EnumerationDiscreteSampler : public DiscreteSampler
 {
 public:
-  EnumerationDiscreteSampler() :
-    DiscreteSampler(), probabilityForUnseenSamples(DEFAULT_PROBABILITY_FOR_UNSEEN_SAMPLES)
+  EnumerationDiscreteSampler()
+    : DiscreteSampler(), probabilityForUnseenSamples(DEFAULT_PROBABILITY_FOR_UNSEEN_SAMPLES)
   {
   }
 
   EnumerationDiscreteSampler(size_t numElements, double probabilityForUnseenSamples =
-      DEFAULT_PROBABILITY_FOR_UNSEEN_SAMPLES) :
-    DiscreteSampler(), probabilityForUnseenSamples(probabilityForUnseenSamples)
+      DEFAULT_PROBABILITY_FOR_UNSEEN_SAMPLES)
+    : DiscreteSampler(), probabilityForUnseenSamples(probabilityForUnseenSamples)
   {
     probabilities = new DenseDoubleVector(denseDoubleVectorClass(
         positiveIntegerEnumerationEnumeration), numElements, 1.0 / (double)numElements);
   }
 
   EnumerationDiscreteSampler(DenseDoubleVectorPtr& probabilities,
-      double probabilityForUnseenSamples = DEFAULT_PROBABILITY_FOR_UNSEEN_SAMPLES) :
-    DiscreteSampler(), probabilityForUnseenSamples(probabilityForUnseenSamples)
+      double probabilityForUnseenSamples = DEFAULT_PROBABILITY_FOR_UNSEEN_SAMPLES)
+    : DiscreteSampler(), probabilityForUnseenSamples(probabilityForUnseenSamples)
   {
     ClassPtr actionClass = denseDoubleVectorClass(positiveIntegerEnumerationEnumeration);
     this->probabilities = new DenseDoubleVector(actionClass, probabilities->getValues());
   }
 
-  EnumerationDiscreteSampler(const EnumerationDiscreteSampler& sampler) :
-    DiscreteSampler(), probabilityForUnseenSamples(sampler.probabilityForUnseenSamples)
+  EnumerationDiscreteSampler(const EnumerationDiscreteSampler& sampler)
+    : DiscreteSampler(), probabilityForUnseenSamples(sampler.probabilityForUnseenSamples)
   {
     probabilities = new DenseDoubleVector(denseDoubleVectorClass(
         positiveIntegerEnumerationEnumeration), sampler.probabilities->getValues());
@@ -78,7 +78,7 @@ public:
     DenseDoubleVectorPtr empiricalFrequencies = new DenseDoubleVector(denseDoubleVectorClass(
         positiveIntegerEnumerationEnumeration), probabilities->getNumElements(), 0.0);
     double increment = 1.0 / (double)dataset.size();
-    for (int i = 0; i < dataset.size(); i++)
+    for (size_t i = 0; i < dataset.size(); i++)
     {
       size_t index = dataset[i].first.getInteger();
       empiricalFrequencies->setValue(index, empiricalFrequencies->getValue(index) + increment);
@@ -86,7 +86,7 @@ public:
 
     // Update frequencies
     double totalNorm = 0;
-    for (int i = 0; i < probabilities->getNumElements(); i++)
+    for (size_t i = 0; i < probabilities->getNumElements(); i++)
     {
       probabilities->setValue(i, juce::jmax(probabilityForUnseenSamples, probabilities->getValue(i)
           + LEARNING_RATE * (empiricalFrequencies->getValue(i) - probabilities->getValue(i))));
@@ -94,7 +94,7 @@ public:
     }
 
     double normalize = 1.0 / totalNorm;
-    for (int i = 0; i < probabilities->getNumElements(); i++)
+    for (size_t i = 0; i < probabilities->getNumElements(); i++)
       probabilities->setValue(i, probabilities->getValue(i) * normalize);
   }
